Separate rename ops
